+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Add auto-registration of `Dictionary` implementations as actual dictionary for users of Sf >= 3.3
+
 ## [2.0.1] - 2017-12-15
 
 ### Fixed